Simplify code by moving PKCE entirely into the AuthorizationCodeFlow class. Add documentation.
